# Drop unnecessary stream operation, reading handles escaping